+++ packages/spacecat-shared-data-access/src/models/site/config.js
@@ -31,6 +31,14 @@ export const configSchema = Joi.object({
       Joi.object({
         disabled: Joi.boolean().optional(),
         excludedURLs: Joi.array().items(Joi.string()).optional(),
+        manualOverwrites: Joi.array().items(Joi.object({
+          brokenTargetURL: Joi.string().optional(),
+          targetURL: Joi.string().optional(),
+        })).optional(),
+        fixedURLs: Joi.array().items(Joi.object({
+          brokenTargetURL: Joi.string().optional(),
+          targetURL: Joi.string().optional(),
+        })).optional(),
       }).unknown(true),
     ).unknown(true),
   }).unknown(true),